+++ client/timesync-client.js
@@ -1,24 +1,30 @@
-//IE8 doesn't have Date.now()
-Date.now = Date.now || function() { return +new Date; };
+import { Meteor } from 'meteor/meteor';
+import { Tracker } from 'meteor/tracker';
+import { HTTP } from 'meteor/http';
 
-TimeSync = {
-  loggingEnabled: true
+// IE8 doesn't have Date.now()
+Date.now = Date.now || function () {
+  return +new Date;
 };
 
-function log(/* arguments */) {
+TimeSync = {
+  loggingEnabled: Meteor.isDevelopment
+};
+
+function log( /* arguments */ ) {
   if (TimeSync.loggingEnabled) {
     Meteor._debug.apply(this, arguments);
   }
 }
 
-var defaultInterval = 1000;
+const defaultInterval = 1000;
 
 // Internal values, exported for testing
 SyncInternals = {
   offset: undefined,
   roundTripTime: undefined,
-  offsetDep: new Deps.Dependency(),
-  syncDep: new Deps.Dependency(),
+  offsetDep: new Tracker.Dependency(),
+  syncDep: new Tracker.Dependency(),
   isSynced: false,
   timeTick: {},
   getDiscrepancy: function (lastTime, currentTime, interval) {
@@ -26,10 +32,10 @@ SyncInternals = {
   }
 };
 
-SyncInternals.timeTick[defaultInterval] = new Deps.Dependency();
+SyncInternals.timeTick[defaultInterval] = new Tracker.Dependency();
 
-var maxAttempts = 5;
-var attempts = 0;
+let maxAttempts = 5;
+let attempts = 0;
 
 /*
   This is an approximation of
@@ -39,39 +45,39 @@ var attempts = 0;
   we should try taking multiple measurements.
  */
 
-var syncUrl;
+let syncUrl;
 if (Meteor.isCordova || Meteor.isDesktop) {
   // Only use Meteor.absoluteUrl for Cordova and Desktop; see
   // https://github.com/meteor/meteor/issues/4696
   // https://github.com/mizzao/meteor-timesync/issues/30
   // Cordova should never be running out of a subdirectory...
-  syncUrl = Meteor.absoluteUrl("_timesync");
-}
-else {
+  syncUrl = Meteor.absoluteUrl('_timesync');
+} else {
   // Support Meteor running in relative paths, based on computed root url prefix
   // https://github.com/mizzao/meteor-timesync/pull/40
   const basePath = __meteor_runtime_config__.ROOT_URL_PATH_PREFIX || '';
-  syncUrl = basePath + "/_timesync";
+  syncUrl = basePath + '/_timesync';
 }
 
-var updateOffset = function() {
-  var t0 = Date.now();
+const updateOffset = function () {
+  const t0 = Date.now();
 
-  HTTP.get(syncUrl, function(err, response) {
-    var t3 = Date.now(); // Grab this now
+  HTTP.get(syncUrl, function (err, response) {
+    const t3 = Date.now(); // Grab this now
     if (err) {
-      //  We'll still use our last computed offset if is defined
-      log("Error syncing to server time: ", err);
-      if (++attempts <= maxAttempts)
+      // We'll still use our last computed offset if is defined
+      log('Error syncing to server time: ', err);
+      if (++attempts <= maxAttempts) {
         Meteor.setTimeout(TimeSync.resync, 1000);
-      else
-        log("Max number of time sync attempts reached. Giving up.");
+      } else {
+        log('Max number of time sync attempts reached. Giving up.');
+      }
       return;
     }
 
     attempts = 0; // It worked
 
-    var ts = parseInt(response.content);
+    const ts = parseInt(response.content, 10);
     SyncInternals.isSynced = true;
     SyncInternals.offset = Math.round(((ts - t0) + (ts - t3)) / 2);
     SyncInternals.roundTripTime = t3 - t0; // - (ts - ts) which is 0
@@ -80,10 +86,10 @@ var updateOffset = function() {
 };
 
 // Reactive variable for server time that updates every second.
-TimeSync.serverTime = function(clientTime, interval) {
+TimeSync.serverTime = function (clientTime, interval) {
   check(interval, Match.Optional(Match.Integer));
   // If a client time is provided, we don't need to depend on the tick.
-  if ( !clientTime ) getTickDependency(interval || defaultInterval).depend();
+  if (!clientTime) getTickDependency(interval || defaultInterval).depend();
 
   SyncInternals.offsetDep.depend(); // depend on offset to enable reactivity
   // Convert Date argument to epoch as necessary
@@ -91,24 +97,24 @@ TimeSync.serverTime = function(clientTime, interval) {
 };
 
 // Reactive variable for the difference between server and client time.
-TimeSync.serverOffset = function() {
+TimeSync.serverOffset = function () {
   SyncInternals.offsetDep.depend();
   return SyncInternals.offset;
 };
 
-TimeSync.roundTripTime = function() {
+TimeSync.roundTripTime = function () {
   SyncInternals.offsetDep.depend();
   return SyncInternals.roundTripTime;
 };
 
-TimeSync.isSynced = function() {
+TimeSync.isSynced = function () {
   SyncInternals.offsetDep.depend();
   return SyncInternals.isSynced;
 };
 
-var resyncIntervalId = null;
+let resyncIntervalId = null;
 
-TimeSync.resync = function() {
+TimeSync.resync = function () {
   if (resyncIntervalId !== null) Meteor.clearInterval(resyncIntervalId);
   updateOffset();
   resyncIntervalId = Meteor.setInterval(updateOffset, 600000);
@@ -116,11 +122,11 @@ TimeSync.resync = function() {
 
 // Run this as soon as we load, even before Meteor.startup()
 // Run again whenever we reconnect after losing connection
-var wasConnected = false;
+let wasConnected = false;
 
-Deps.autorun(function() {
-  var connected = Meteor.status().connected;
-  if ( connected && !wasConnected ) TimeSync.resync();
+Tracker.autorun(function () {
+  const connected = Meteor.status().connected;
+  if (connected && !wasConnected) TimeSync.resync();
   wasConnected = connected;
 });
 
@@ -128,17 +134,17 @@ Deps.autorun(function() {
 // somewhat lenient, or a CPU-intensive operation can trigger a re-sync even
 // when the offset is still accurate. In any case, we're not going to be able to
 // catch very small system-initiated NTP adjustments with this, anyway.
-var tickCheckTolerance = 5000;
+const tickCheckTolerance = 5000;
 
-var lastClientTime = Date.now();
+let lastClientTime = Date.now();
 
 // Set up a new interval for any amount of reactivity.
 function getTickDependency(interval) {
 
-  if ( !SyncInternals.timeTick[interval] ) {
-    var dep  = new Deps.Dependency();
+  if (!SyncInternals.timeTick[interval]) {
+    let dep = new Tracker.Dependency();
 
-    Meteor.setInterval(function() {
+    Meteor.setInterval(function () {
       dep.changed();
     }, interval);
 
@@ -149,17 +155,17 @@ function getTickDependency(interval) {
 }
 
 // Set up special interval for the default tick, which also watches for re-sync
-Meteor.setInterval(function() {
-  var currentClientTime = Date.now();
+Meteor.setInterval(function () {
+  let currentClientTime = Date.now();
 
-  var discrepancy = SyncInternals.getDiscrepancy(lastClientTime, currentClientTime, defaultInterval);
+  let discrepancy = SyncInternals.getDiscrepancy(lastClientTime, currentClientTime, defaultInterval);
   if (Math.abs(discrepancy) < tickCheckTolerance) {
     // No problem here, just keep ticking along
     SyncInternals.timeTick[defaultInterval].changed();
   } else {
     // resync on major client clock changes
     // based on http://stackoverflow.com/a/3367542/1656818
-    log("Clock discrepancy detected. Attempting re-sync.");
+    log('Clock discrepancy detected. Attempting re-sync.');
     // Refuse to compute server time and try to guess new server offset. Guessing only works if the server time hasn't changed.
     SyncInternals.offset = SyncInternals.offset - discrepancy;
     SyncInternals.isSynced = false;