+++ package.js
@@ -1,12 +1,12 @@
 Package.describe({
-  name: "mizzao:timesync",
-  summary: "NTP-style time synchronization between server and client",
-  version: "0.5.1",
-  git: "https://github.com/mizzao/meteor-timesync.git"
+  name: 'mizzao:timesync',
+  summary: 'NTP-style time synchronization between server and client',
+  version: '0.6.0',
+  git: 'https://github.com/mizzao/meteor-timesync.git'
 });
 
 Package.onUse(function (api) {
-  api.versionsFrom("1.2.0.1");
+  api.versionsFrom('METEOR@1.3');
 
   api.use([
     'check',
@@ -14,27 +14,30 @@ Package.onUse(function (api) {
     'http'
   ], 'client');
 
-  api.use('webapp', 'server');
+  api.use(['webapp'], 'server');
 
-  api.use('ecmascript');
+  api.use(['ecmascript']);
 
   // Our files
-  api.addFiles('timesync-server.js', 'server');
-  api.addFiles('timesync-client.js', 'client');
+  api.addFiles('server/index.js', 'server');
+  api.addFiles('client/index.js', 'client');
 
   api.export('TimeSync', 'client');
-  api.export('SyncInternals', 'client', {testOnly: true} );
+  api.export('SyncInternals', 'client', {
+    testOnly: true
+  });
 });
 
 Package.onTest(function (api) {
   api.use([
+    'ecmascript',
     'tinytest',
     'test-helpers'
   ]);
 
-  api.use(["tracker", "underscore"], 'client');
+  api.use(['tracker'], 'client');
 
-  api.use("mizzao:timesync");
+  api.use('mizzao:timesync');
 
   api.addFiles('tests/client.js', 'client');
 });