+++ tests/client.js
@@ -1,4 +1,4 @@
-Tinytest.add("timesync - tick check - normal tick", function(test) {
+Tinytest.add("timesync - tick check - normal tick", function (test) {
   var lastTime = 5000;
   var currentTime = 6000;
   var interval = 1000;
@@ -6,7 +6,7 @@ Tinytest.add("timesync - tick check - normal tick", function(test) {
   test.equal(SyncInternals.getDiscrepancy(lastTime, currentTime, interval), 0);
 });
 
-Tinytest.add("timesync - tick check - slightly off", function(test) {
+Tinytest.add("timesync - tick check - slightly off", function (test) {
   var lastTime = 5000;
   var currentTime = 6500;
   var interval = 1000;
@@ -18,7 +18,7 @@ Tinytest.add("timesync - tick check - slightly off", function(test) {
   test.equal(SyncInternals.getDiscrepancy(lastTime, currentTime, interval), -500);
 });
 
-Tinytest.add("timesync - tick check - big jump", function(test) {
+Tinytest.add("timesync - tick check - big jump", function (test) {
   var lastTime = 5000;
   var currentTime = 0;
   var interval = 1000;
@@ -34,7 +34,7 @@ Tinytest.add("timesync - tick check - big jump", function(test) {
   TODO: add tests for proper dependencies in reactive functions
  */
 
-Tinytest.addAsync("timesync - basic - initial sync", function(test, next) {
+Tinytest.addAsync("timesync - basic - initial sync", function (test, next) {
 
   function success() {
     var syncedTime = TimeSync.serverTime();
@@ -46,7 +46,7 @@ Tinytest.addAsync("timesync - basic - initial sync", function(test, next) {
     // always be true in PhantomJS tests where client/server are the same
     // machine, although it might fail in development environments, for example
     // when the server and client are different VMs.
-    test.isTrue( Math.abs(syncedTime - Date.now()) < 1000 );
+    test.isTrue(Math.abs(syncedTime - Date.now()) < 1000);
 
     next();
   }
@@ -59,24 +59,26 @@ Tinytest.addAsync("timesync - basic - initial sync", function(test, next) {
   simplePoll(TimeSync.isSynced, success, fail, 5000, 100);
 });
 
-Tinytest.addAsync("timesync - basic - serverTime format", function(test, next) {
+Tinytest.addAsync("timesync - basic - serverTime format", function (test, next) {
 
-  test.isTrue(_.isNumber( TimeSync.serverTime() ));
+  test.isTrue(_.isNumber(TimeSync.serverTime()));
 
-  test.isTrue(_.isNumber( TimeSync.serverTime(null) ));
+  test.isTrue(_.isNumber(TimeSync.serverTime(null)));
 
   // Accept Date as client time
-  test.isTrue(_.isNumber( TimeSync.serverTime(new Date()) ));
+  test.isTrue(_.isNumber(TimeSync.serverTime(new Date())));
 
   // Accept epoch as client time
-  test.isTrue(_.isNumber( TimeSync.serverTime(Date.now()) ));
+  test.isTrue(_.isNumber(TimeSync.serverTime(Date.now())));
 
   next();
 });
 
-Tinytest.addAsync("timesync - basic - different sync intervals", function(test, next) {
+Tinytest.addAsync("timesync - basic - different sync intervals", function (test, next) {
 
-  var aCount = 0, bCount = 0, cCount = 0;
+  var aCount = 0,
+    bCount = 0,
+    cCount = 0;
 
   var a = Tracker.autorun(function () {
     TimeSync.serverTime(null, 500);
@@ -95,7 +97,7 @@ Tinytest.addAsync("timesync - basic - different sync intervals", function(test, 
 
   var testInterval = 4990;
 
-  Meteor.setTimeout(function() {
+  Meteor.setTimeout(function () {
 
     test.equal(aCount, 10); // 0, 500, 1000, 1500 ...
     // not going to be 5 since the first tick won't generate this dep